project/requirements.txt
# server
fastapi[all]==0.74.1
jinja2==3.1.1

# code formatting
flake8==4.0.1
black==22.3.0
# database
aerich==0.6.3
redis==4.2.2
tortoise-orm[asyncpg]==0.19.0

# unit/integration testing

pytest==7.1.1
pytest-asyncio==0.18.3
httpx==0.22.0
asgi-lifespan==1.0.1
asynctest==0.13.0

# password hashing/auth
passlib[bycrypt]==1.7.4
python-jose[cryptography]==3.3.0
# pyjwt

# AWS
boto3==1.21.42

fastapi-mail==1.1.4